FileAPI/line-endings/window_toNativeLineEndings.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>File API: window_toNativeLineEndings</title>
    <link rel="author" title="Intel" href="http://www.intel.com">
    <link rel="help" href="http://www.w3.org/TR/FileAPI/#methodsOnLineEndings">
    <link rel="help" href="http://www.w3.org/TR/FileAPI/#convenienceAPI">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <meta name="assert" content="Check the implementation of the toNativeLineEndings in FileAPI">
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(function () {
        assert_true("toNativeLineEndings" in window, "window has toNativeLineEndings method");
      }, "Check if window has toNativeLineEndings method");
    </script>
  </body>
</html>
